use getStaticPaths in the dynamic pages
